+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules/
 .idea/
-sdk/dist
+dist/
