+++ README.md
@@ -1,12 +1,7 @@
 # SybilCenter SDK
 
-This is a monorepo for SybilCenter SDK. In one line you can issue a Verifiable Credential proving that your user owns a social media account. Supported are Twitter, Discord, and GitHub.
-
-The repository includes SDK as well as a Next.js application, showing how to use the SDK in your application.
-
-We require that in your application, an Etehreum provider conforming to [EIP-1193](https://eips.ethereum.org/EIPS/eip-1193) is available, like MetaMask, or Tokenary. The provider is used to confirm that a subject of a credential, i.e. an entity we issue the credential about, is genuine.
-
-Let's assume that the provider (`window.ethereum` for MetaMask, for example) is available, and the user allowed your application to access it (`window.ethereum.enable()` for MetaMask). Then in the application code you add the following lines, for example to prove your user have a Twitter account:
+In one line you can issue a Verifiable Credential proving that your user owns a social media account.
+Supported are Twitter, Discord, and GitHub.
 
 ```typescript
 import { Sybil, EthRequestSigner, type IEIP1193Provider } from "@sybil-center/sdk";
@@ -20,23 +15,31 @@ const signer new EthRequestSigner(injected);
 const credential = await sybil.credential('twitter-account', signer.sign) // This returns a Verifiable Credential
 ```
 
-Please, consult [the SDK readme](./sdk/README.md) for more technical information, and [demo-app README](./demo-app/README.md) about how to run a demo locally, and where to look for an example of code in React.
+## Install
 
-## How to use the repository
-
-If you intend to use the repository we require [`pnpm` package manager](http://pnpm.io).
-
-The following commands would install all the necessary dependencies and build JS code:
-
-```
-pnpm install
-pnpm run build
+```shell
+npm install @sybil-center/sdk # Feel free to use pnpm or yarn
 ```
 
-If you would like to see the demo on your machine, switch to `demo-app` folder, and start the application:
-```
-cd demo-app
-pnpm run dev
-```
+## Supported credentials
 
-Then open [`http://localhost:3000`](http://localhost:3000) in a browser with an Ethereum wallet set up.
+Supported designators for `sybil.credential` call:
+
+- `twitter-account`, available fields:
+  - `id: string` - permanent unique Twitter account identifier
+  - `username: string` - current Twitter account username
+- `github-account`, available fields:
+  - `id: number` - permanent unique GitHub account identifier
+  - `username: string` - current GitHub account username aka `login`
+- `discord-account`, available fields:
+  - `id: string` - permanent unique Discord account identifier
+  - `username: string` - the user's username, not unique across the Discord platform
+  - `discriminator: string` - the user's 4-digit discord-tag
+
+If you call `sybil.credential` method with any of the credential designators, type of a resulting credential is provided for you by TypeScript. To get, for example, Twitter username from `credential` as in excerpt above, you would use `credential.credentialSubject.twitter.username`.
+
+# Demo App
+
+The repository also contains demo React/Next.js application, showing how to use the SDK in your application.
+
+Please, consult [demo-app README](./demo-app/README.md) about how to run a demo locally, and where to look for an example of code in React.
