--- pnpm-workspace.yaml
@@ -1,3 +1,0 @@
-packages:
-  - sdk
-  - demo-app
