--- sdk/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@sybil-center/sdk",
-  "version": "0.0.2",
-  "description": "Request a Verifiable Credential from the web",
-  "license": "(Apache-2.0 AND MIT)",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sybil-center/sybil.git"
-  },
-  "type": "module",
-  "files": [
-    "dist/"
-  ],
-  "sideEffects": false,
-  "typesVersions": {
-    "*": {
-      "*": [
-        "dist/*"
-      ]
-    }
-  },
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "build": "../node_modules/.bin/tsc -p ."
-  },
-  "dependencies": {
-    "uint8arrays": "^4.0.3"
-  }
-}
