--- sdk/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "./src/",
-    "sourceRoot": "./src/",
-    "outDir": "./dist/"
-  },
-  "exclude": ["./dist/"]
-}
