+++ src/clients/eth.client.ts
@@ -1,11 +1,12 @@
-import {
-  EthOwnershipOptions,
-  EthAccOwnershipProvider,
-  type IEthAccountOwnershipVC
-} from "../providers/eth-acc-ownership.provider.js";
-import type { HttpClient } from "../util/http-client.js";
-import type { SignFn } from "../util/sign-fn.type.js";
 import type { IClient } from "./client.type.js";
+import {
+  EthAccOwnershipProvider,
+  EthOwnershipOptions,
+  IEthAccountOwnershipVC
+} from "../providers/eth-acc-ownership.provider.js";
+import { HttpClient } from "../util/http-client.js";
+import type { SignFn } from "../util/sign-fn.type.js";
+
 
 export class EthAccountClient implements IClient<IEthAccountOwnershipVC, EthOwnershipOptions> {
   private readonly provider: EthAccOwnershipProvider;
