+++ src/clients/github-account.client.ts
@@ -1,13 +1,13 @@
 import type { IClient } from "./client.type.js";
-import type { HttpClient } from "../util/http-client.js";
-import type { SignFn } from "../util/sign-fn.type.js";
-import { repeatUntil } from "../util/repeat-until.js";
 import {
   GitHubAccOwnershipProvider,
   type GitHubAccOwnershipVC,
-  GitHubOwnershipOptions
+  type GitHubOwnershipOptions
 } from "../providers/github-acc-ownership.provider.js";
+import { HttpClient } from "../util/http-client.js";
+import type { SignFn } from "../util/sign-fn.type.js";
 import { popupFeatures } from "../util/view.js";
+import { repeatUntil } from "../util/repeat-until.js";
 
 export class GithubAccountClient
   implements IClient<GitHubAccOwnershipVC, GitHubOwnershipOptions> {
