+++ src/clients/twitter-account.client.ts
@@ -1,35 +1,34 @@
 import type { IClient } from "./client.type.js";
 import {
-  DiscordAccOwnershipProvider,
-  type DiscordAccOwnVC,
-  DiscordOwnershipOptions
-} from "../providers/discord-acc-ownership.provider.js";
-import type { HttpClient } from "../util/http-client.js";
+  type ITwitterAccountOwnershipVC,
+  TwitterAccOwnershipProvider,
+  type TwitterOwnershipOptions
+} from "../providers/twitter-acc-ownership.provider.js";
+import { HttpClient } from "../util/http-client.js";
 import type { SignFn } from "../util/sign-fn.type.js";
-import { repeatUntil } from "../util/repeat-until.js";
 import { popupFeatures } from "../util/view.js";
+import { repeatUntil } from "../util/repeat-until.js";
 
-export class DiscordAccountClient
-  implements IClient<DiscordAccOwnVC, DiscordOwnershipOptions> {
+export class TwitterAccountClient implements IClient<ITwitterAccountOwnershipVC, TwitterOwnershipOptions> {
 
   constructor(
     httpClient: HttpClient,
-    private readonly provider = new DiscordAccOwnershipProvider(httpClient)
+    private readonly provider = new TwitterAccOwnershipProvider(httpClient)
   ) {}
 
   async issueCredential(
     signFn: SignFn,
-    opt?: DiscordOwnershipOptions
-  ): Promise<DiscordAccOwnVC> {
+    opt?: TwitterOwnershipOptions
+  ): Promise<ITwitterAccountOwnershipVC> {
     const payload = await this.provider.getPayload({
       redirectUrl: opt?.redirectUrl
     });
     const popup = window.open(
       payload.authUrl,
       "_blank",
-      opt?.windowFeature ? opt?.windowFeature : popupFeatures()
+      opt?.windowFeatures ? opt?.windowFeatures : popupFeatures()
     );
-    if (!popup) throw new Error(`Can not open popup window to authenticate in Discord`);
+    if (!popup) throw new Error(`Can not open popup window to authenticate in Twitter`);
     await repeatUntil<boolean>(
       (r) => r,
       1000,
