+++ src/sybil.ts
@@ -1,13 +1,13 @@
-import { IClient } from "./clients/client.type.js";
-import { TwitterAccountClient } from "./clients/twitter-account-client.js";
-import { SignFn } from "./util/sign-fn.type.js";
-import { ITwitterAccountOwnershipVC, TwitterOwnershipOptions } from "./providers/twitter-acc-ownership.provider.js";
-import { DiscordAccOwnVC, DiscordOwnershipOptions } from "./providers/discord-acc-ownership.provider.js";
-import { DiscordAccountClient } from "./clients/discord-account-client.js";
-import { EthOwnershipOptions, IEthAccountOwnershipVC } from "./providers/eth-acc-ownership.provider.js";
-import { EthAccountClient } from "./clients/eth-account-client.js";
-import { GitHubOwnershipOptions, GitHubAccOwnershipVC } from "./providers/github-acc-ownership.provider.js";
-import { GithubAccountClient } from "./clients/github-account-client.js";
+import type { IClient } from "./clients/client.type.js";
+import { TwitterAccountClient } from "./clients/twitter-account.client.js";
+import type { SignFn } from "./util/sign-fn.type.js";
+import { type ITwitterAccountOwnershipVC, type TwitterOwnershipOptions } from "./providers/twitter-acc-ownership.provider.js";
+import { type DiscordAccOwnVC, type DiscordOwnershipOptions } from "./providers/discord-acc-ownership.provider.js";
+import { DiscordAccountClient } from "./clients/discord-account.client.js";
+import { type EthOwnershipOptions, type IEthAccountOwnershipVC } from "./providers/eth-acc-ownership.provider.js";
+import { EthAccountClient } from "./clients/eth-account.client.js";
+import { type GitHubOwnershipOptions, type GitHubAccOwnershipVC } from "./providers/github-acc-ownership.provider.js";
+import { GithubAccountClient } from "./clients/github-account.client.js";
 import { HttpClient } from "./util/http-client.js";
 
 export { ITwitterAccountOwnershipVC, DiscordAccOwnVC, IEthAccountOwnershipVC, GitHubAccOwnershipVC };
