+++ src/util/proof-type.ts
